transcription_service/requirements.txt
setuptools-rust
watchdog
transformers
torch
torchaudio
python-dotenv
boto3
Flask